backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==3.0.2
mysql==0.0.3
mysqlclient==2.2.4
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
tomli==2.0.2
typing_extensions==4.12.2
uvicorn==0.32.0
virtualenv==20.27.1
